remove ERC165Utils, move UniversalResolver to match interface PR, change Lookup.extended to bits, start UR2 hardhat tests
